Move rider functions in :drive RideCache/Service to :ride module files
