+++ feature/drive/src/commonMain/kotlin/local/RideCache.kt
@@ -4,17 +4,13 @@
 
 package org.pointyware.xyz.drive.local
 
-import org.pointyware.xyz.core.entities.ride.Ride
 import org.pointyware.xyz.core.entities.Uuid
-import org.pointyware.xyz.drive.data.RideSearchResult
+import org.pointyware.xyz.core.entities.ride.Ride
 
 /**
  *
  */
 interface RideCache {
-    suspend fun saveDestinations(query: String, searchResult: RideSearchResult)
-    suspend fun getDestinations(query: String): RideSearchResult?
-    suspend fun dropDestinations(query: String)
     suspend fun saveRide(ride: Ride)
     suspend fun dropRide(ride: Ride)
 }
@@ -22,20 +18,6 @@ interface RideCache {
 class RideCacheImpl(
 
 ): RideCache {
-
-    private val queryCache = mutableMapOf<String, RideSearchResult>()
-
-    override suspend fun saveDestinations(query: String, searchResult: RideSearchResult) {
-        queryCache[query] = searchResult
-    }
-
-    override suspend fun getDestinations(query: String): RideSearchResult? {
-        return queryCache[query]
-    }
-
-    override suspend fun dropDestinations(query: String) {
-        queryCache.remove(query)
-    }
 
     private val rideCache = mutableMapOf<Uuid, Ride>()
